requirement.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
Django==5.1
django-allauth==64.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.7
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
requests==2.32.3
rest-framework-simplejwt==0.0.2
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
